Backend-API/requirements.txt
Flask==2.0.3
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Marshmallow==0.14.0
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.9.3

